fix(ci): isolate credential-backed E2E smokes
