+++ test/e2e/mock-parity.json
@@ -11,6 +11,10 @@
       ]
     },
     {
+      "live": "test/e2e/live/inference-routing-provider-smoke.test.ts",
+      "liveOnlyReason": "Requires real provider credentials to prove provider responses and sandbox credential isolation"
+    },
+    {
       "live": "test/e2e/live/full-e2e.test.ts",
       "fast": ["test/e2e/support/onboard-performance.test.ts"]
     },
