+++ tools/e2e/risk-signal.ts
@@ -12,8 +12,6 @@ export type E2eRiskSignal = {
   passed: number;
   failed: number;
   skipped: number;
-  // Additive version-1 field; older PR heads omit it and are treated as zero.
-  optionalSkipped?: number;
   pending: number;
   unhandledErrors: number;
   runReason: "passed" | "failed" | "interrupted";
